MSLargeFileUploadTask: Catch exception when no data is available
When the internet connection is lost during a large file upload the
completion handler of the uploadTask is invoked. However, the first
thing the handler does is converting the data passed to a JSON. If the
data is nil or empty, NSJSONSerialization throws an exception.
This patch handles that gracefully and simply calls the completion
handler initially passed to createUploadSessionFromRequest